fix: fix code style
